% Auto-generated: do not edit by hand
\name{vegaVega}

\alias{vegaVega}

\title{Vega component}

\description{
在src/lib/fragments/Vega.react.js文件修改
}

\usage{
vegaVega(id=NULL, clickData=NULL, height=NULL, hoverData=NULL,
spec=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{clickData}{Named list. }

\item{height}{Character. }

\item{hoverData}{Named list. }

\item{spec}{Named list. A label that will be printed when this component is rendered.}

\item{width}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

